Reduce unbalanced errors in space allocation
Before this change, a group of six layout objects spread as follows:

* Share:  5  -  1  -  1  -  1  -  1  -  5
* Min:    0  - 38  - 15  - 15  -  9  -  0

Given available space of 382 would spread as follows:

* Space: 136 - 38  - 26  - 26  - 26  - 130

Note that the first and last *should* have the same space allocated,
but do not; the larger than 1/14 share minimum on the second layout
object caused an imbalance.  Now this is handled properly and becomes:

* Space: 133 - 38  - 26  - 26  - 26  - 133

Now the first and last have properly split the 10/13 share of what
remains after satisfying the second object's minimum (38).