Fix behaviour with no lines and empty lines
Both are corner cases we have to take into account.
Previously these cases couldn't occur because of a (wrong) fix in
`Span.lines` that prevented to ever return an empty span list.
But that bad fix is now gone and we have to be able to cope with
empty span lists now.