Fix off-by-one error in set-x/y-scroll
It's feels odd to have the *entire* visible content scroll away, so
make sure that the last line/cell remain visible.  Yes, with ragged
text you could end up scrolling to the lower right corner and not
being able to see anything (because the very last cell was empty),
but I'm not trying to fight that particular battle right now.